+++ src/common/expression_type.cpp
@@ -1,5 +1,7 @@
 #include "src/common/include/expression_type.h"
 
+#include <stdexcept>
+
 namespace graphflow {
 namespace common {
