+++ src/common/include/operations/comparison_operations.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cassert>
+#include <functional>
 
 #include "src/common/include/types.h"
 #include "src/common/include/value.h"
