+++ src/expression/BUILD.bazel
@@ -1,25 +1,40 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
 
+
 cc_library(
-    name = "expression",
+    name = "logical_expression",
     srcs = [
         "logical/logical_expression.cpp",
         "logical/logical_literal_expression.cpp",
-        "physical/physical_binary_expression.cpp",
-        "physical/physical_expression.cpp",
-        "physical/physical_unary_expression.cpp",
     ],
     hdrs = [
         "include/logical/logical_expression.h",
         "include/logical/logical_literal_expression.h",
         "include/logical/logical_node_expression.h",
         "include/logical/logical_rel_expression.h",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/common:expression_type",
+        "//src/common:types",
+    ],
+)
+
+cc_library(
+    name = "physical_expression",
+    srcs = [
+        "physical/physical_binary_expression.cpp",
+        "physical/physical_expression.cpp",
+        "physical/physical_unary_expression.cpp",
+    ],
+    hdrs = [
         "include/physical/physical_binary_expression.h",
         "include/physical/physical_expression.h",
         "include/physical/physical_unary_expression.h",
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//src/common",
+        "//src/common:data_chunk",
+        "//src/common:vector_operations",
     ],
 )
