+++ src/loader/include/utils.h
@@ -8,6 +8,7 @@
 #include "src/common/include/compression_scheme.h"
 #include "src/common/include/configs.h"
 #include "src/common/include/types.h"
+#include "src/common/include/utils.h"
 #include "src/storage/include/data_structure/lists/list_headers.h"
 #include "src/storage/include/data_structure/lists/lists_metadata.h"
 #include "src/storage/include/data_structure/utils.h"
