+++ src/processor/physical_plan/operator/load_csv/load_csv.cpp
@@ -2,6 +2,8 @@
 
 #include <cassert>
 
+#include "src/common/include/string.h"
+
 using namespace graphflow::common;
 
 namespace graphflow {
