+++ src/storage/include/data_structure/column.h
@@ -2,6 +2,7 @@
 
 #include <string.h>
 
+#include "src/common/include/string.h"
 #include "src/common/include/types.h"
 #include "src/common/include/vector/node_vector.h"
 #include "src/storage/include/data_structure/data_structure.h"
