+++ src/storage/include/store/rels_store.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "src/common/include/types.h"
 #include "src/storage/include/catalog.h"
 #include "src/storage/include/data_structure/column.h"
 #include "src/storage/include/data_structure/lists/lists.h"
