+++ test/common/vector/operations/vector_hash_nodeid_operations_test.cpp
@@ -3,18 +3,18 @@
 #include "src/common/include/data_chunk/data_chunk.h"
 #include "src/common/include/operations/hash_operations.h"
 #include "src/common/include/vector/node_vector.h"
+#include "src/common/include/vector/operations/vector_node_id_operations.h"
 
 using namespace graphflow::common;
 using namespace std;
 
-/*
 TEST(VectorHashNodeIDTests, nonSequenceNodeIDTest) {
     auto dataChunk = make_shared<DataChunk>();
     dataChunk->state->size = 1000;
     dataChunk->state->numSelectedValues = 1000;
 
     NodeIDCompressionScheme compressionScheme;
-    auto nodeVector = make_shared<NodeIDVector>(100, compressionScheme);
+    auto nodeVector = make_shared<NodeIDVector>(100, compressionScheme, false);
     dataChunk->append(nodeVector);
     auto nodeData = (uint64_t*)nodeVector->values;
 
@@ -27,8 +27,7 @@ TEST(VectorHashNodeIDTests, nonSequenceNodeIDTest) {
         nodeData[i] = i * 10 + 1;
     }
 
-    auto hashNodeIDOp = ValueVector::getUnaryOperation(HASH_NODE_ID);
-    hashNodeIDOp(*nodeVector, *result);
+    VectorNodeIDOperations::Hash(*nodeVector, *result);
     for (int32_t i = 0; i < 1000; i++) {
         auto expected = operation::murmurhash64(i * 10 + 1) ^ operation::murmurhash64(100);
         ASSERT_EQ(resultData[i], expected);
@@ -36,7 +35,7 @@ TEST(VectorHashNodeIDTests, nonSequenceNodeIDTest) {
 
     // Set dataChunk to flat
     dataChunk->state->currPos = 8;
-    hashNodeIDOp(*nodeVector, *result);
+    VectorNodeIDOperations::Hash(*nodeVector, *result);
     auto expected = operation::murmurhash64(8 * 10 + 1) ^ operation::murmurhash64(100);
     auto pos = result->state->getCurrSelectedValuesPos();
     ASSERT_EQ(resultData[pos], expected);
@@ -48,7 +47,8 @@ TEST(VectorHashNodeIDTests, sequenceNodeIDTest) {
     dataChunk->state->numSelectedValues = 1000;
 
     label_t commonLabel = 100;
-    auto nodeVector = make_shared<NodeIDSequenceVector>(commonLabel);
+    NodeIDCompressionScheme compressionScheme;
+    auto nodeVector = make_shared<NodeIDVector>(commonLabel, compressionScheme, true);
     nodeVector->setStartOffset(10);
     dataChunk->append(nodeVector);
 
@@ -56,8 +56,7 @@ TEST(VectorHashNodeIDTests, sequenceNodeIDTest) {
     dataChunk->append(result);
     auto resultData = (uint64_t*)result->values;
 
-    auto hashNodeIDOp = ValueVector::getUnaryOperation(HASH_NODE_ID);
-    hashNodeIDOp(*nodeVector, *result);
+    VectorNodeIDOperations::Hash(*nodeVector, *result);
     for (int32_t i = 0; i < 1000; i++) {
         auto expected = operation::murmurhash64(10 + i) ^ operation::murmurhash64(100);
         ASSERT_EQ(resultData[i], expected);
@@ -65,8 +64,8 @@ TEST(VectorHashNodeIDTests, sequenceNodeIDTest) {
 
     // Set dataChunk to flat
     dataChunk->state->currPos = 8;
-    hashNodeIDOp(*nodeVector, *result);
+    VectorNodeIDOperations::Hash(*nodeVector, *result);
     auto expected = operation::murmurhash64(10 + 8) ^ operation::murmurhash64(100);
     auto pos = result->state->getCurrSelectedValuesPos();
     ASSERT_EQ(resultData[pos], expected);
-}*/
+}
